[fix] optmize webhook
